refactor(core): migrate middleware to proxy pattern (Next.js 16)
Next.js 16 deprecates middleware.ts in favor of proxy.ts. This renames
the entry point, export, composition utilities, and directory to align
with the new convention.

Also enforces Node 24 via engines field in package.json, since proxy
runs on the Node.js runtime (not Edge).
