+++ core/proxies/with-channel-id.ts
@@ -1,8 +1,8 @@
 import { getChannelIdFromLocale } from '~/channels.config';
 
-import { type MiddlewareFactory } from './compose-middlewares';
+import { type ProxyFactory } from './compose-proxies';
 
-export const withChannelId: MiddlewareFactory = (next) => {
+export const withChannelId: ProxyFactory = (next) => {
   return (request, event) => {
     const locale = request.headers.get('x-bc-locale') ?? '';
     const channelId = getChannelIdFromLocale(locale) ?? '';
